1.0.0-SNAPSHOT

+++ pdf-generator/src/main/java/cz/sysnet/pdf/generator/PdfFactory.java
@@ -1,6 +1,7 @@
 package cz.sysnet.pdf.generator;
 
 import java.io.BufferedReader;
+import java.io.BufferedWriter;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
@@ -10,7 +11,11 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -32,6 +37,14 @@ import net.sf.jasperreports.engine.xml.JRXmlLoader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+/**
+ * PdfFactory je singleton, ktery slouzi ke generovani PDF dokumentu z dodanych dat pomoci reportovaci sablony Jaser Reports
+ * Obsahuje radu produkcnich a pomocnych metod, se kterymi tvori uceleny balik.
+ * 
+ * @author 		Radim Jäger
+ * @version		1.0.0
+ * 
+ */
 public class PdfFactory {
 	private static volatile PdfFactory instance;
 	private static Object mutex = new Object();
@@ -53,7 +66,8 @@ public class PdfFactory {
 	public static final String DATA_PATH = (System.getProperty("user.home") + FILE_SEPARATOR + "pdf_factory" + FILE_SEPARATOR).replace(FILE_SEPARATOR+FILE_SEPARATOR, FILE_SEPARATOR);
 	public static final String TEMPLATE_PATH = DATA_PATH + "templates"  + FILE_SEPARATOR;
 	public static final String PDF_PATH = DATA_PATH + "pdf"  + FILE_SEPARATOR;
-	
+	public static final String CONFIG_PATH = DATA_PATH;
+	public static final String CONFIG_TEMPLATES_PATH = CONFIG_PATH + "templates.json";
 	
 	private PdfFactory() {
 		super();
@@ -81,13 +95,26 @@ public class PdfFactory {
 			}
 		}
 		return result;
-	}
+	}	
 	
+	
+	/**
+	 * Pomocná metoda. Slouží k převodu souboru do vstupního streamu
+	 * 
+	 * @param filePath	Úplná cesta k souboru 
+	 * @return			Vrací InputStream, v případě chyby null 
+	 */
 	public static InputStream fileToInputStream(String filePath) {
 		File file = new File(filePath);
 		return fileToInputStream(file);
 	}
 	
+	/**
+	 * Pomocná metoda. Slouží k převodu souboru do vstupního streamu
+	 * 
+	 * @param file		Vstupní soubor 
+	 * @return			Vrací InputStream, v případě chyby null 
+	 */
 	public static InputStream fileToInputStream(File file) {
 		try {
 			InputStream is = new FileInputStream(file);
@@ -100,13 +127,27 @@ public class PdfFactory {
 		}
 	}
 	
+	/**
+	 * Pomocná metoda. Slouží k převodu souboru JSON do vstupního streamu bytového pole. 
+	 *  
+	 * @param jsonFilePath	Úplná cesta k souboru JSON
+	 * @return				Vrací ByteArrayInputStream, v případě chyby null
+	 */
 	public static ByteArrayInputStream jsonFileToByteArrayInputStream(String jsonFilePath) {
 		File jsonFile = new File(jsonFilePath);
 		return jsonFileToByteArrayInputStream(jsonFile);
 	}
 	
+	/**
+	 * Pomocná metoda. Slouží k převodu souboru JSON do vstupního streamu bytového pole. 
+	 *  
+	 * @param jsonFile		Soubor JSON
+	 * @return				Vrací ByteArrayInputStream, v případě chyby null
+	 */
 	public static ByteArrayInputStream jsonFileToByteArrayInputStream(File jsonFile) {
 		try {
+			if (jsonFile == null) throw new PdfGeneratorException("JSON file == null");
+			if (!jsonFile.exists()) throw new PdfGeneratorException("JSON file does nots exist: " + jsonFile.getCanonicalPath());
 			BufferedReader br = new BufferedReader(new FileReader(jsonFile));
 			JsonObject jo = jsonParser.parse(br).getAsJsonObject();
 			String js = gson.toJson(jo);
@@ -119,6 +160,12 @@ public class PdfFactory {
 		}
 	}
 	
+	/**
+	 * Pomocná metoda. Slouží k převodu dat JSON do vstupního streamu bytového pole. 
+	 *  
+	 * @param string		Data JSON
+	 * @return				Vrací ByteArrayInputStream, v případě chyby null
+	 */
 	public static ByteArrayInputStream stringToByteArrayInputStream(String string) {
 		try {
 			byte[] utf8JsonString = string.getBytes("UTF8");
@@ -132,14 +179,29 @@ public class PdfFactory {
 		}
 	}
 	
-	public String addTemplate(String key, String sourceFile) {
+	/**
+	 * Do mapy šablon přidá šablonu. Šablona se překopíruje ze zdrojové cesty do standardního umístení šablon
+	 *  
+	 * @param key			klíč šablony
+	 * @param sourceFile	Úplná cesta ke zdrojovému souboru šablony
+	 * @param removeSource	Odstraní zdrojový soubor
+	 * @return				Vrací cestu, kam byla šablona umístěna
+	 */
+	public String addTemplate(String key, String sourceFile, boolean removeSource) {
 		try {
 			File file = new File(sourceFile);
-			String targetFile = TEMPLATE_PATH + "key" + FILE_SEPARATOR + file.getName();
+			String targetFile = TEMPLATE_PATH + key + FILE_SEPARATOR + file.getName();
 			Path sourceFilePath = Paths.get(sourceFile);
+			if (!Files.exists(sourceFilePath)) throw new PdfGeneratorException("Zdrojový soubor neexistuje: " + sourceFilePath.toString());
 			Path targetFilePath = Paths.get(targetFile);
+			if (!Files.exists(targetFilePath)) {
+				Path td = Files.createDirectories(targetFilePath);
+				LOG.info("Directory created: " + td.toString());
+			}
 			Files.copy(sourceFilePath, targetFilePath, StandardCopyOption.REPLACE_EXISTING);
 			this.templateMap.put(key, targetFile);
+			if (removeSource) Files.deleteIfExists(sourceFilePath);
+			LOG.info("Template added: " + key +": " + targetFilePath.toString());
 			return targetFilePath.toString();
 			
 		} catch (Exception e) {
@@ -149,13 +211,28 @@ public class PdfFactory {
 		}
 	}
 	
-	public String addTemplateResource(String key, String sourceFile) {
+	/**
+	 * Přidá jeden soubor k šabloně s klíčem (například obrázek). Soubor se překopíruje ze zdrojové cesty do standardního umístení šablon 
+	 * 
+	 * @param key			klíč šablony
+	 * @param sourceFile	Úplná cesta ke zdrojovému souboru
+	 * @param removeSource	Odstraní zdrojový soubor
+	 * @return				Vrací cestu, kam byl soubor umístěn
+	 */
+	public String addTemplateResource(String key, String sourceFile, boolean removeSource) {
 		try {
 			File file = new File(sourceFile);
-			String targetFile = TEMPLATE_PATH + "key" + FILE_SEPARATOR + file.getName();
+			String targetFile = TEMPLATE_PATH + key + FILE_SEPARATOR + file.getName();
 			Path sourceFilePath = Paths.get(sourceFile);
+			if (!Files.exists(sourceFilePath)) throw new PdfGeneratorException("Zdrojový soubor neexistuje: " + sourceFilePath.toString());
 			Path targetFilePath = Paths.get(targetFile);
+			if (!Files.exists(targetFilePath)) {
+				Path td = Files.createDirectories(targetFilePath);
+				LOG.info("Directory created: " + td.toString());
+			}
 			Files.copy(sourceFilePath, targetFilePath, StandardCopyOption.REPLACE_EXISTING);
+			if (removeSource) Files.deleteIfExists(sourceFilePath);
+			LOG.info("Template resource added: " + key +": " + targetFilePath.toString());
 			return targetFilePath.toString();
 			
 		} catch (Exception e) {
@@ -165,21 +242,96 @@ public class PdfFactory {
 		}
 	}
 	
-	public String addTemplateResource(String key, List<String> sourceFileList) {
+	/**
+	 * Přidá seznam souborů k šabloně s klíčem (například obrázky). Soubory se překopírují ze zdrojové cesty do standardního umístění šablon 
+	 * 
+	 * @param key			klíč šablony
+	 * @param sourceFileList	Seznam úplných cest ke zdrojovým souborům
+	 * @param removeSource	Odstraní zdrojové soubory
+	 * @return				Vrací JSON seznam cest, kam byly soubory umístěny
+	 */
+	public String addTemplateResource(String key, List<String> sourceFileList, boolean removeSource) {
 		List<String> out = new ArrayList<String>(); 
 		for (String sourceFile:sourceFileList) {
-			out.add(this.addTemplateResource(key, sourceFile));
+			out.add(this.addTemplateResource(key, sourceFile, removeSource));
 		}
 		return gson.toJson(out);
 	}
 	
+	/**
+	 * Vytvoří PDF dokument implicitního jména na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param data			Vstupní data jako String
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
+	public String createPdfFromJsonByKey(String key, String data) {
+		return createPdfFromJsonByKey(key, data, null);		
+	}
 	
+	/**
+	 * Vytvoří PDF dokument na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param data			Vstupní data jako String
+	 * @param pdfFileName	Úplná cesta pro umístění PDF
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
+	public String createPdfFromJsonByKey(String key, String data, String pdfFileName) {
+		return createPdfFromJsonByKey(key, stringToByteArrayInputStream(data), pdfFileName);		
+	}
+	
+	/**
+	 * Vytvoří PDF dokument implicitního jména na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param jsonData		Soubor se vstupními daty
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
+	public String createPdfFromJsonByKey(String key, File jsonData) {
+		return createPdfFromJsonByKey(key, jsonData, null);		
+	}
+	
+	
+	/**
+	 * Vytvoří PDF dokument na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param jsonData		Soubor se vstupními daty
+	 * @param pdfFileName	Úplná cesta pro umístění PDF
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
+	public String createPdfFromJsonByKey(String key, File jsonData, String pdfFileName) {
+		return createPdfFromJsonByKey(key, jsonFileToByteArrayInputStream(jsonData), pdfFileName);		
+	}
+	
+	/**
+	 * Vytvoří PDF dokument implicitního jména na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param dataStream	Stream vstupních dat
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
+	public String createPdfFromJsonByKey(String key, ByteArrayInputStream dataStream) {
+		return createPdfFromJsonByKey(key, dataStream, null);
+	}
+	
+	/**
+	 * Vytvoří PDF dokument na základě šablony dané klíčem a vstupního datového streamu
+	 * 
+	 * @param key			klíč šablony
+	 * @param dataStream	Stream vstupních dat
+	 * @param pdfFileName	Úplná cesta pro umístění PDF
+	 * @return				Vrací cestu k vygenerovanému PDF nebo null v případě chyby
+	 */
 	public String createPdfFromJsonByKey(String key, ByteArrayInputStream dataStream, String pdfFileName) {
 		try {
 			if (this.templateMap == null) throw new PdfGeneratorException("Template map does not esist");
 			if (this.templateMap.isEmpty()) throw new PdfGeneratorException("Template map is empty");
 			if (!this.templateMap.containsKey(key)) throw new PdfGeneratorException("Template map does not contain template '"+key+"'");
-			String templateFilePath = this.templateMap.get(key); 
+			String templateFilePath = this.templateMap.get(key);
+			if (pdfFileName == null) pdfFileName = generatePdfFileName(key.toLowerCase());
+			if (pdfFileName.isEmpty()) pdfFileName = generatePdfFileName(key.toLowerCase());
 			return createPdfFromJson(templateFilePath, dataStream, null, pdfFileName);
 			
 		} catch (Exception e) {
@@ -234,6 +386,15 @@ public class PdfFactory {
 		return createPdfFromJson(templateStream, dataStream, null, pdfFileName);
 	}
 	
+	/**
+	 * Vytvoří PDF dokument na základě šablony dané streamem, vstupního datového streamu, mapy parametrů a názvu výstupního souboru
+	 * 
+	 * @param templateStream	Vstupní stream šablony Jasper Reports (jrxml) 
+	 * @param dataStream		ByteArrayInputStream vstuní data
+	 * @param parameters		parametry nastavení šablony
+	 * @param pdfFileName		Název výstupního souboru PDF (jen název, ne cesta) 
+	 * @return					Vrací cestu k uloženému výstupnímu souboru nebo null při chybě. 
+	 */
 	public String createPdfFromJson(InputStream templateStream, ByteArrayInputStream dataStream, Map<String, Object> parameters, String pdfFileName) {
 		try {
 			JasperDesign design = JRXmlLoader.load(templateStream);
@@ -244,7 +405,16 @@ public class PdfFactory {
 			    parameters.put("title", "SYSNET PDF Generator");
 		    }
 		    JasperPrint jasperPrint = JasperFillManager.fillReport(report, parameters, ds);
+		    if (pdfFileName == null) pdfFileName = generatePdfFileName();
+		    if (pdfFileName.isEmpty()) pdfFileName = generatePdfFileName();
 		    String pdfPath = PDF_PATH + pdfFileName;
+			Path targetFilePath = Paths.get(PDF_PATH);
+			if (!Files.exists(targetFilePath)) {
+				Path td = Files.createDirectories(targetFilePath);
+				LOG.info("Directory created: " + td.toString());
+			}
+			targetFilePath = Paths.get(pdfPath);
+			Files.deleteIfExists(targetFilePath);
 		    JasperExportManager.exportReportToPdfFile(jasperPrint, pdfPath);
 		    LOG.info("PdfFactory.createPdfFromJson - PDF created: " +  pdfPath);
 		    return pdfPath;
@@ -265,8 +435,71 @@ public class PdfFactory {
 		this.templateMap = templateMap;
 	}
 	
+	public boolean storeTemplateMap() {
+		try {
+			Path path = Paths.get(CONFIG_TEMPLATES_PATH);
+			Files.deleteIfExists(path);
+			BufferedWriter bw = Files.newBufferedWriter(path);
+			gson.toJson(this.templateMap, bw);
+			bw.flush();
+			bw.close();
+			LOG.info("Template map stored");
+			return true;
+			
+		} catch (Exception e) {
+			LOG.error("PdfFactory.storeTemplateMap: " + e.getMessage());
+			e.printStackTrace();
+			return false;
+		}
+	}
 	
+	public void cleanTemplateMap() {
+		this.templateMap = new HashMap<String, String>();
+	}
 	
+	@SuppressWarnings("unchecked")
+	public boolean loadTemplateMap() {
+		try {
+			if (this.templateMap == null) this.templateMap = new HashMap<String, String>();
+			Path path = Paths.get(CONFIG_TEMPLATES_PATH);
+			if (!Files.exists(path)) throw new PdfGeneratorException("Konfigurace šablon neexistuje: " + path.toString());
+			BufferedReader br = Files.newBufferedReader(path);
+			if (br == null) throw new PdfGeneratorException("BufferedReader == null");
+			Map<String, String> out = null;
+	        out = gson.fromJson(br, this.templateMap.getClass());
+	        this.templateMap = out;
+			LOG.info("Template map loaded");
+			return true;
+			
+		} catch (Exception e) {
+			LOG.error("PdfFactory.loadTemplateMap: " + e.getMessage());
+			e.printStackTrace();
+			return false;
+		}
+	}
 	
+	public static String generatePdfFileName() {
+		return generatePdfFileName(null);		
+	}
+	
+	public static String generatePdfFileName(String key) {
+		try {
+			if (key == null) key = "document";
+			if (key.isEmpty()) key = "document";
+			
+			DateFormat df = new SimpleDateFormat("yyyyMMddHHmmssSSS");
+			//DateFormat df = new SimpleDateFormat("MM/dd/yyyy HH:mm:ss");
+			Date today = Calendar.getInstance().getTime();    
+			String datePart = df.format(today);
+			
+			String out = key + "_" + datePart + ".pdf";
+			return out;
+			
+		} catch (Exception e) {
+			LOG.error("PdfFactory.generatePdfFileName: " + e.getMessage());
+			e.printStackTrace();
+			return null;
+		}
+	}
 	
 }
